StyleSheet correction
