+++ todolist/app/components/Footer/styles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  footer: {
+    backgroundColor: 'white',
+    flexDirection: 'column',
+    height: 50
+  },
+  footerButton: {
+    alignItems: 'center',
+    backgroundColor: 'white',
+    height: 50,
+    justifyContent: 'center'
+  },
+  textFooterButton: {
+    color: '#CB4335',
+    fontSize: 14,
+    fontWeight: "bold"
+  }
+});
+
+export default styles;
