+++ todolist/app/components/List/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    paddingTop: 10
+  }
+});
+
+export default styles;
