+++ todolist/app/components/Title/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    backgroundColor: '#87CEEB',
+    flexDirection: 'column',
+    height: 50,
+    justifyContent: 'center'
+  },
+  titleText: {
+    color: "#FFF",
+    fontSize: 18,
+    fontWeight: "bold"
+  }
+});
+
+export default styles;
